websocketclient stuff in the index
