+++ src/index.ts
@@ -1,21 +1,35 @@
 import * as schedule from "node-schedule";
-import config from '../config';
 import Tradingbot from './tradingbot';
+import CronHelper from './helpers/cronHelper';
+import WebSocketService from './binance/websocket';
+import { WebsocketClient, WsKey } from './../node_modules/binance/lib/websocket-client';
+import txtLogger from './helpers/txt-logger';
+import { LogLevel } from './models/log-level'; // todo aram check if loglevel can be included in txtLogger class
+import WebSocket from 'isomorphic-ws';
 
 // app.listen(app.get("port"), () => {
 //     console.log(("App is running"), app.get("env"));
     console.log("App is running");
-    const configInterval: string = config.orderConditions[0].interval;
-    const tradingBotInterval = configInterval.slice(0, -1);
-    const cronExpression = '*/' + tradingBotInterval + ' * * * *';
 
-    console.log('cronExpression = ' + cronExpression);
-    
-    schedule.scheduleJob(cronExpression, async function () {
-        var tradingBot = new Tradingbot();
-        await tradingBot.runProgram();
+    // setup
+    const cronExpression = CronHelper.GetCronExpression();
+    const wsService: WebSocketService = new WebSocketService();
+    let websocketClient: WebsocketClient = wsService.generateWebsocketClient();
+
+    websocketClient.subscribeSpotUserDataStream();
+
+    websocketClient.on('open', async (data: {
+        wsKey: WsKey;
+        ws: WebSocket;
+        event?: any;
+    }) => {
+        txtLogger.writeToLogFile(`Websocket event - connection opened open:', ${data.wsKey}, ${data.ws.url}`);
+
+        schedule.scheduleJob(cronExpression, async function () {
+            var tradingBot = new Tradingbot(wsService, websocketClient);
+            await tradingBot.runProgram();
+        });
     });
-// });
 
 // app.on('close', () => {
 //     console.log("App is Closing");
